Implement floating point vector registers & lower vector insert/extract.
x/y/zmm should be represented as vectors of floats(/doubles), so that LLVM can optimize loops over the elemnts.
(Unroled) Loops get created when we copy values from one vector to another, or zero out the enclosing registers implicitly.

LLVM is smart enough to automatically cast f32 vectors to f64 vectors with half the number of elements, because everything is properly aligned.