txlat: handle symbols of size 0 in the ahead-of-time pass
Some function symbols in the function table have a size of 0, which means they
were previously ignored. With this commit, we translate chunk starting from the
base address of these symbols of size 0 until we reach the next address
corresponding to a symbol. The translated chunk might not fully correspond to
actual code (there might be some data between the two symbols), but we don't
really care as this "junk" translation will never be reached anyways.