[txlat] Emit RELA relocations corresponding to those in the binary
They are inserted directly into the .rela.dyn section of the result binary bypassing the assembler and linker, so they can't interfere and possibly alter the relocations.

Relocations that involve a symbol therefore have to be patched after the fact to point to the correct entry in the new dynsym table.

This is done by parsing the generated ELF and iterating over the relocations. All the ones with any of the 4 highest bits of the reloc type set need to have a transform applied.