[txlat] [doc] Support generating libraries.
It will automatically detect if the input was a shared library and generate a library if that is the case.
This can then be used with the -l option when generating the respective executable.
Libraries are only checked with the exact path given and *not* any default or otherwise specified search paths.
Library dependencies are not checked and will most likely cause errors in the linker if missing.