[txlat] Add support for IFUNCs and IRELATIVE relocations.
We can't run the actual resolvers before the runtime is initialized, but the dynamic linker would run them when resolving relocations.

So we generate a wrapper which calls the resolver the first time it is invoked and saves the result (similar to how a PLT works).

We can't just resolve IRELATIVE relocations at runtime initialization because IFUNC symbols and the respective IRELATIVE need to give the same address.