[RV64] Fix no-op nodes causing reads of outdated register values.
Some nodes are effectively no-op (like bit inserts into a 128 bit value that only change one of the halfs). Previously those just directly returned the input register. If the output of the no-op is used by multiple nodes this means the input register was now used by multiple nodes. But read_reg nodes of GPR among others require their output to be moved to a fresh register, if it is used by multiple nodes. So we add a move here, should get optimized in code gen in most cases if not actually needed.