[ir] [front] Add operation_value port to atomic nodes.
It holds the result of the operation, i.e. the value ending up in the memory location at the completion of the operation.

The value port is now the value at the memory location prior to the modification.
This is typically what RMW operations place in the out register.

This simplifies flag generation in the backend as flags typically need both result and the inputs.

Xadd is now simply an atomic add which uses the redefined value port.
Xchg also now uses the value port instead of doing "magic" register modifications.