Remove cpu_state from return values
We now no longer return the cpu_state, as the caller always has a
reference to the CPU state. This change requires using the latest
version of our LLVM fork, as that skips one return register for the
Arancini calling convention.

This should free up one register for certain optimizations, if the cpu
state is not accessed from one point in the function anymore.

There is an issue with the call to the internal_main_loop. If LLVM
is allowed to tail-optimize it, something is broken and the translated
binary crashes. For now, we just set the call to not tail-optimize. At
some point in the future, we should fix this.